renaming classes, applying new concepts to prepare for testing
