+++ TastyScript/Lang/Functions/FunctionConnectDevice.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using TastyScript.Lang.Exceptions;
 using TastyScript.Lang.Extensions;
-using TastyScript.Lang.Token;
+using TastyScript.Lang.Tokens;
 
 namespace TastyScript.Lang.Functions
 {
