+++ TastyScript/Lang/Token/IToken.cs
@@ -3,9 +3,23 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.Serialization;
 using TastyScript.Lang.Extensions;
+using TastyScript.Lang.TokenOLD;
 
-namespace TastyScript.Lang.Token
+namespace TastyScript.Lang.Tokens
 {
+    internal class Token
+    {
+        public string Name { get; private set; }
+        public string Value { get; private set; }
+        public Token(string name, string val)
+        {
+            Name = name;
+            Value = val;
+        }
+    }
+    //
+    //OLD VVVV
+    //
     [Serializable]
     internal abstract class Token<T> : IToken<T>
     {
