+++ TastyScript/Lang/Token/TParameter.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using TastyScript.Lang.Tokens;
 
-namespace TastyScript.Lang.Token
+namespace TastyScript.Lang.TokenOLD
 {
     [Serializable]
     internal class TParameter : Token<List<IBaseToken>>
